Move all the fetch requests and their urls into cart, user and products JS files.
