+++ public/Scripts/Amazon_checkout.js
@@ -1,20 +1,7 @@
 import {getLoginStatus, sumCartItems, shipping, findProductInCartTable, postProductToCart} from './user.js'
 import {getProducts} from './products.js';
-const productsTableUrl = 'https://api.backendless.com/059E0E6C-3A70-434F-B0EE-230A6650EEAE/3AB37559-1318-4AAE-8B26-856956A63050/data/products';
-let cartUrl = 'https://api.backendless.com/059E0E6C-3A70-434F-B0EE-230A6650EEAE/3AB37559-1318-4AAE-8B26-856956A63050/data/cart'; //we should put the Id of that specific record which match the conditions we set before in the findProductInCartTable() at the end of this URL if we wanna update.
-
-async function deleteItemInCart(userInfo, productObjectId) {
-  const itemRecord = await findProductInCartTable(0, userInfo['userId'], productObjectId, userInfo['userToken']);
-  const response = await fetch(`${cartUrl}/${itemRecord[0]['objectId']}`, {
-    method: 'DELETE',
-    headers: {
-      'user-token': userInfo['userToken']
-    }
-  }) 
-  if(response.ok) {
-    alert('You have deleted this item!');
-  }
-}
+import { deleteItemInCart } from './cart.js';
+import { productsTableUrl } from '.products.js';
 
 function activateUpdateButton() {   // here only activate the update buttons, not click them. The click event only happen in the eventlistener when the update buttons are clicked.
   const updateButton = document.querySelectorAll(".cart-product-update-button");
