+++ public/Scripts/products.js
@@ -1,4 +1,4 @@
-const productsTableUrl = 'https://api.backendless.com/059E0E6C-3A70-434F-B0EE-230A6650EEAE/3AB37559-1318-4AAE-8B26-856956A63050/data/products';
+export const productsTableUrl = 'https://api.backendless.com/059E0E6C-3A70-434F-B0EE-230A6650EEAE/3AB37559-1318-4AAE-8B26-856956A63050/data/products';
 const userLoginUrl = 'https://supplekick-us.backendless.app/api/users/login';
 const email = '433lll433@gmail.com';
 const password = 'Shirahama';
